Split CLI help and telemetry responsibilities
Amp-Thread-ID: https://ampcode.com/threads/T-01a0128d-e4fc-77fa-8be7-0d4ab85c64ef
