+++ internal/cli/cli_test.go
@@ -75,278 +75,6 @@ func canonicalTempDir(t *testing.T) string {
 	return dir
 }
 
-func TestRunHelpAndVersion(t *testing.T) {
-	tests := []struct {
-		name       string
-		args       []string
-		wantOutput string
-	}{
-		{name: "help flag", args: []string{"--help"}, wantOutput: "validate"},
-		{name: "help command", args: []string{"help"}, wantOutput: "run-job"},
-		{name: "command help", args: []string{"help", "compile"}, wantOutput: "buildkite-gha compile --event-path"},
-		{name: "upload help", args: []string{"help", "upload"}, wantOutput: "--runner-queue"},
-		{name: "upload help flag", args: []string{"upload", "--help"}, wantOutput: "--runner-queue"},
-		{name: "command help flag", args: []string{"run-job", "--help"}, wantOutput: "--plan <path>"},
-		{name: "version flag", args: []string{"--version"}, wantOutput: "buildkite-gha test-version\n"},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			var stdout, stderr bytes.Buffer
-			if code := Run(test.args, &stdout, &stderr, "test-version"); code != 0 {
-				t.Fatalf("Run() code = %d, want 0; stderr = %q", code, stderr.String())
-			}
-			if !strings.Contains(stdout.String(), test.wantOutput) {
-				t.Errorf("Run() stdout = %q, want it to contain %q", stdout.String(), test.wantOutput)
-			}
-			if stderr.Len() != 0 {
-				t.Errorf("Run() stderr = %q, want empty", stderr.String())
-			}
-		})
-	}
-}
-
-func TestUploadHelpFormsMatch(t *testing.T) {
-	outputs := make([]string, 0, 2)
-	for _, args := range [][]string{{"help", "upload"}, {"upload", "--help"}} {
-		var stdout, stderr bytes.Buffer
-		if code := Run(args, &stdout, &stderr, "test-version"); code != 0 {
-			t.Fatalf("Run(%q) code = %d, stderr = %q", args, code, stderr.String())
-		}
-		outputs = append(outputs, stdout.String())
-	}
-	if outputs[0] != outputs[1] || !strings.Contains(outputs[0], "explicit .yml or .yaml path") || !strings.Contains(outputs[0], "failed or skipped workflows become top-level replacement steps") || !strings.Contains(outputs[0], "--runner-image") || !strings.Contains(outputs[0], "every scheduled workflow group is eligible") {
-		t.Fatalf("upload help outputs differ or omit runner profile or aggregate workflow options:\nhelp command: %q\nhelp flag: %q", outputs[0], outputs[1])
-	}
-}
-
-func TestPluginIsHiddenAndZeroArgument(t *testing.T) {
-	for _, args := range [][]string{{"--help"}, {"help"}} {
-		var stdout, stderr bytes.Buffer
-		if code := Run(args, &stdout, &stderr, "test-version"); code != 0 {
-			t.Fatalf("Run(%q) code = %d, stderr = %q", args, code, stderr.String())
-		}
-		if strings.Contains(stdout.String(), "plugin") {
-			t.Fatalf("Run(%q) exposed plugin command: %q", args, stdout.String())
-		}
-	}
-	for _, args := range [][]string{{"help", "plugin"}, {"plugin", "--help"}} {
-		var stdout, stderr bytes.Buffer
-		if code := Run(args, &stdout, &stderr, "test-version"); code != 2 {
-			t.Fatalf("Run(%q) code = %d, want 2", args, code)
-		}
-		if stdout.Len() != 0 || strings.Contains(stderr.String(), "buildkite-gha plugin") {
-			t.Fatalf("Run(%q) exposed plugin help: stdout=%q stderr=%q", args, stdout.String(), stderr.String())
-		}
-	}
-}
-
-func TestCommandCompletionTelemetryPlacementAndExitSemantics(t *testing.T) {
-	type event struct {
-		Event      string `json:"event"`
-		Properties struct {
-			Command       string `json:"command"`
-			Outcome       string `json:"outcome"`
-			ClientVersion string `json:"client_version"`
-			DurationMS    int64  `json:"duration_ms"`
-			FailurePhase  string `json:"failure_phase"`
-			FailureCode   string `json:"failure_code"`
-		} `json:"properties"`
-	}
-	events := make(chan event, 2)
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if got := r.Header.Get("Authorization"); got != "Token telemetry-token" {
-			t.Errorf("Authorization = %q", got)
-		}
-		var received event
-		if err := json.NewDecoder(r.Body).Decode(&received); err != nil {
-			t.Errorf("decode telemetry event: %v", err)
-		}
-		events <- received
-		w.WriteHeader(http.StatusAccepted)
-	}))
-	defer server.Close()
-	t.Setenv("BUILDKITE_AGENT_ENDPOINT", server.URL+"/v3")
-	t.Setenv("BUILDKITE_JOB_ID", cliTestJobID)
-	t.Setenv("BUILDKITE_AGENT_ACCESS_TOKEN", "telemetry-token")
-	t.Setenv("BUILDKITE_GHA_TELEMETRY_DISABLED", "")
-
-	for _, test := range []struct {
-		args        []string
-		wantCode    int
-		wantCommand string
-	}{
-		{args: []string{"plugin", "unexpected"}, wantCode: 2, wantCommand: "plugin_import"},
-		{args: []string{"run-job", "--unknown"}, wantCode: 2, wantCommand: "run_job"},
-	} {
-		var stdout, stderr bytes.Buffer
-		if code := run(test.args, &stdout, &stderr, "test-version", &cliCaptureRunner{}); code != test.wantCode {
-			t.Fatalf("run(%q) code = %d, stderr = %q; want %d", test.args, code, stderr.String(), test.wantCode)
-		}
-		received := <-events
-		if received.Event != "pipelines:buildkite_gha:command_completed" || received.Properties.Command != test.wantCommand || received.Properties.Outcome != "usage_error" || received.Properties.ClientVersion != "test-version" || received.Properties.DurationMS < 0 || received.Properties.FailurePhase != "configuration" || received.Properties.FailureCode != "unknown" {
-			t.Fatalf("run(%q) event = %#v", test.args, received)
-		}
-	}
-}
-
-func TestCommandCompletionTelemetryOptOut(t *testing.T) {
-	requests := 0
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		requests++
-		w.WriteHeader(http.StatusAccepted)
-	}))
-	defer server.Close()
-	t.Setenv("BUILDKITE_AGENT_ENDPOINT", server.URL)
-	t.Setenv("BUILDKITE_JOB_ID", cliTestJobID)
-	t.Setenv("BUILDKITE_AGENT_ACCESS_TOKEN", "telemetry-token")
-	t.Setenv("BUILDKITE_GHA_TELEMETRY_DISABLED", "true")
-
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--unknown"}, &stdout, &stderr, "dev", &cliCaptureRunner{}); code != 2 {
-		t.Fatalf("run() code = %d, stderr = %q; want 2", code, stderr.String())
-	}
-	if requests != 0 {
-		t.Fatalf("telemetry requests = %d, want 0", requests)
-	}
-}
-
-func TestCommandCompletionTelemetryFailureDoesNotChangeExit(t *testing.T) {
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		http.Error(w, "private response", http.StatusInternalServerError)
-	}))
-	defer server.Close()
-	t.Setenv("BUILDKITE_AGENT_ENDPOINT", server.URL)
-	t.Setenv("BUILDKITE_JOB_ID", cliTestJobID)
-	t.Setenv("BUILDKITE_AGENT_ACCESS_TOKEN", "telemetry-token")
-	t.Setenv("BUILDKITE_GHA_TELEMETRY_DISABLED", "")
-
-	var stdout, stderr bytes.Buffer
-	if code := run([]string{"run-job", "--unknown"}, &stdout, &stderr, "dev", &cliCaptureRunner{}); code != 2 {
-		t.Fatalf("run() code = %d, stderr = %q; want 2", code, stderr.String())
-	}
-	if strings.Contains(stderr.String(), "private response") || strings.Contains(stderr.String(), "telemetry") {
-		t.Fatalf("telemetry failure leaked to stderr: %q", stderr.String())
-	}
-}
-
-func TestRevisionedDevelopmentVersionKeepsPlanCompatibilityVersion(t *testing.T) {
-	if got := commandVersion("dev+0123456789ab"); got != "dev" {
-		t.Fatalf("commandVersion() = %q, want dev", got)
-	}
-}
-
-func TestCommandTelemetryDetailsCollectTypedDiagnostics(t *testing.T) {
-	details := &commandTelemetryDetails{}
-	details.observe(compatibility.ProcessingReport{Diagnostics: []compatibility.Diagnostic{
-		{Level: "error", Code: compiler.CodeWorkflowSyntax, Stage: string(compiler.StageWorkflowParsing), Message: "must not be uploaded"},
-		{Level: "error", Code: compiler.CodeWorkflowSyntax, Stage: string(compiler.StageWorkflowParsing), Message: "different sensitive text"},
-		{Level: "warning", Code: "W_ACTION_RUNTIME_UNKNOWN", Stage: string(compiler.StageAdmission), Message: "must not be uploaded"},
-		{Level: "error", Code: "E_FUTURE_UNALLOWLISTED", Stage: string(compiler.StageGraph), Message: "must not be uploaded"},
-	}})
-	got := details.telemetryDetails()
-	if got.FailurePhase != "parsing" || got.FailureCode != "E_WORKFLOW_SYNTAX" {
-		t.Fatalf("failure details = %#v", got)
-	}
-	want := []telemetry.Diagnostic{
-		{Code: compiler.CodeWorkflowSyntax, Severity: telemetry.SeverityError},
-		{Code: "W_ACTION_RUNTIME_UNKNOWN", Severity: telemetry.SeverityWarning},
-	}
-	if !reflect.DeepEqual(got.Diagnostics, want) {
-		t.Fatalf("diagnostics = %#v, want %#v", got.Diagnostics, want)
-	}
-}
-
-func TestUnprovenActionRuntimeIgnoresNativeAdapters(t *testing.T) {
-	bundleWith := func(locks ...plan.ActionLock) compiler.Bundle {
-		return compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-			Actions: locks,
-			Steps:   []plan.Step{{Kind: "uses", Uses: "actions/checkout@v4"}},
-		}}}}
-	}
-	checkout := plan.ActionLock{Source: "github", Repository: "actions/checkout"}
-	upload := plan.ActionLock{Source: "github", Repository: "actions/upload-artifact"}
-	download := plan.ActionLock{Source: "github", Repository: "actions/download-artifact"}
-	setupGo := plan.ActionLock{Source: "github", Repository: "actions/setup-go"}
-	local := plan.ActionLock{Source: "workspace", Path: "actions/build"}
-
-	for _, test := range []struct {
-		name  string
-		locks []plan.ActionLock
-		want  bool
-	}{
-		{name: "no actions"},
-		{name: "native adapters only", locks: []plan.ActionLock{checkout, upload, download}},
-		{name: "public action", locks: []plan.ActionLock{setupGo}, want: true},
-		{name: "local action", locks: []plan.ActionLock{local}, want: true},
-		{name: "native adapter alongside public action", locks: []plan.ActionLock{checkout, setupGo}, want: true},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			if got := bundleRunsUnprovenActions(bundleWith(test.locks...)); got != test.want {
-				t.Fatalf("bundleRunsUnprovenActions() = %t, want %t", got, test.want)
-			}
-		})
-	}
-}
-
-func TestHandledReportErrorsDoNotAttributeCommandFailure(t *testing.T) {
-	details := &commandTelemetryDetails{}
-	details.addReportDiagnostics(compatibility.ProcessingReport{Diagnostics: []compatibility.Diagnostic{
-		{Level: "error", Code: compiler.CodeExpressionInvalid, Stage: string(compiler.StageExpressions), Message: "emitted as a failing step"},
-	}})
-	got := details.forOutcome(telemetry.OutcomeFailure)
-	if got.FailurePhase != telemetry.FailurePhaseUnknown || got.FailureCode != telemetry.FailureCodeUnknown {
-		t.Fatalf("handled error attributed a later failure: %#v", got)
-	}
-	want := []telemetry.Diagnostic{{Code: compiler.CodeExpressionInvalid, Severity: telemetry.SeverityError}}
-	if !reflect.DeepEqual(got.Diagnostics, want) {
-		t.Fatalf("diagnostics = %#v, want %#v", got.Diagnostics, want)
-	}
-}
-
-func TestCommandTelemetryDetailsPreservesFirstFailurePhase(t *testing.T) {
-	details := &commandTelemetryDetails{}
-	details.setFailurePhase(telemetry.FailurePhaseExecution)
-	details.setFailurePhase(telemetry.FailurePhaseResultPublication)
-	got := details.forOutcome(telemetry.OutcomeFailure)
-	if got.FailurePhase != telemetry.FailurePhaseExecution || got.FailureCode != telemetry.FailureCodeUnknown {
-		t.Fatalf("failure details = %#v", got)
-	}
-}
-
-func TestTelemetryOutcomePreservesCommandSemantics(t *testing.T) {
-	for _, test := range []struct {
-		name       string
-		code       int
-		conclusion string
-		contextErr error
-		want       telemetry.Outcome
-	}{
-		{name: "success", conclusion: "success", want: telemetry.OutcomeSuccess},
-		{name: "failure", code: 1, conclusion: "failure", want: telemetry.OutcomeFailure},
-		{name: "result publication failure", code: 1, conclusion: "success", want: telemetry.OutcomeFailure},
-		{name: "cancelled result", code: 1, conclusion: "cancelled", want: telemetry.OutcomeCancelled},
-		{name: "cancelled context", code: 1, contextErr: context.Canceled, want: telemetry.OutcomeCancelled},
-		{name: "skipped", conclusion: "skipped", want: telemetry.OutcomeSkipped},
-		{name: "skipped result publication failure", code: 1, conclusion: "skipped", want: telemetry.OutcomeFailure},
-		{name: "tolerated", code: buildkitepipeline.ContinueOnErrorExitStatus, conclusion: "success", want: telemetry.OutcomeToleratedFailure},
-		{name: "usage", code: 2, want: telemetry.OutcomeUsageError},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			if got := telemetryOutcome(test.code, test.conclusion, test.contextErr); got != test.want {
-				t.Fatalf("telemetryOutcome() = %q, want %q", got, test.want)
-			}
-		})
-	}
-}
-
-func TestRunJobUntypedFailurePhaseIsUnknown(t *testing.T) {
-	details := (&commandTelemetryDetails{}).forOutcome(telemetry.OutcomeFailure)
-	if details.FailurePhase != telemetry.FailurePhaseUnknown || details.FailureCode != telemetry.FailureCodeUnknown {
-		t.Fatalf("run-job fallback details = %#v", details)
-	}
-}
-
 func TestPluginTelemetryReportsUnprovenActionRuntime(t *testing.T) {
 	requireImporterHost(t)
 	repository := writeUploadWorkflowRepository(t, map[string]string{
@@ -7504,40 +7232,6 @@ func (r *cliCaptureRunner) Run(ctx context.Context, dir, name string, args []str
 }
 
 var _ transport.Runner = (*cliCaptureRunner)(nil)
-
-func TestRunUsageErrors(t *testing.T) {
-	tests := []struct {
-		name string
-		args []string
-		want string
-	}{
-		{name: "missing command", want: "Usage:"},
-		{name: "unknown command", args: []string{"nope"}, want: `unknown command "nope"`},
-		{name: "unknown help command", args: []string{"help", "nope"}, want: `unknown command "nope"`},
-		{name: "version arguments", args: []string{"--version", "extra"}, want: "does not accept arguments"},
-		{name: "upload outside Buildkite", args: []string{"upload", "--runtime-queue", "hosted", "workflow.yml"}, want: "BUILDKITE=true"},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			if len(test.args) != 0 && test.args[0] == "upload" {
-				requireImporterHost(t)
-			}
-			t.Setenv("BUILDKITE", "")
-			t.Setenv("BUILDKITE_STEP_KEY", "")
-			var stdout, stderr bytes.Buffer
-			if code := Run(test.args, &stdout, &stderr, "dev"); code != 2 {
-				t.Fatalf("Run() code = %d, want 2", code)
-			}
-			if stdout.Len() != 0 {
-				t.Errorf("Run() stdout = %q, want empty", stdout.String())
-			}
-			if !strings.Contains(stderr.String(), test.want) {
-				t.Errorf("Run() stderr = %q, want it to contain %q", stderr.String(), test.want)
-			}
-		})
-	}
-}
 
 func TestRunJobExecutesBoundPlanAndWritesResult(t *testing.T) {
 	workspace := t.TempDir()
